[Fix] #1- 코드리뷰 반영
