--- Winey/Winey/Source/Global/Extension/Date+.swift
@@ -1,19 +1,0 @@
-//
-//  Date+.swift
-//  Winey
-//
-//  Created by 김인영 on 2023/07/05.
-//
-
-import UIKit
-import Foundation
-
-extension Date {
-    
-    func getTodayDateToString(_ format: String) -> String {
-        var dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = format
-        let today = dateFormatter.string(from: Date())
-        return today
-    }
-}
