--- Winey/Winey/Source/Global/Extension/String+.swift
@@ -1,74 +1,0 @@
-//
-//  String+.swift
-//  Winey
-//
-//  Created by 김인영 on 2023/07/05.
-//
-
-import UIKit
-
-extension String {
-
-    func setAttributeString(range: NSRange, font: UIFont, textColor: UIColor) -> NSAttributedString {
-        let attributeString = NSMutableAttributedString(string: self)
-        attributeString.addAttribute(.foregroundColor, value: textColor, range: range)
-        attributeString.addAttribute(.font, value: font, range: range)
-        return attributeString
-    }
-    
-    /// String을 UIImage로 반환하는 메서드
-    func makeImage() -> UIImage? {
-        if let data = Data(base64Encoded: self, options: .ignoreUnknownCharacters) {
-            return UIImage(data: data)
-        }
-        return nil
-    }
-    
-    /// 서버에서 들어온 Date String을 Date 타입으로 반환하는 메서드
-    private func toDate() -> Date {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"
-        dateFormatter.timeZone = TimeZone(identifier: "KST")
-        if let date = dateFormatter.date(from: self) {
-            return date
-        } else {
-            print("toDate() convert error")
-            return Date()
-        }
-    }
-    
-    /// serverTimeToString의 용도 정의
-    enum TimeStringCase {
-        case forNotification
-        case forDefault
-    }
-    
-    /// 서버에서 들어온 Date String을 UI에 적용 가능한 String 타입으로 반환하는 메서드
-    func serverTimeToString(forUse: TimeStringCase) -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateFormat = "yy/MM/dd"
-        
-        let currentTime = Int(Date().timeIntervalSince1970)
-        
-        switch forUse {
-        case .forNotification:
-            let getTime = self.toDate().timeIntervalSince1970
-            let displaySec = currentTime - Int(getTime)
-            let displayMin = displaySec / 60
-            let displayHour = displayMin / 60
-            let displayDay = displayHour / 24
-            
-            if displayDay >= 1 {
-                return dateFormatter.string(from: self.toDate())
-            } else if displayHour >= 1 {
-                return "\(displayHour)시간 전"
-            } else if displayMin >= 1 {
-                return "\(displayMin)분 전"
-            } else {
-                return "1분 전"
-            }
-        case .forDefault:
-            return dateFormatter.string(from: self.toDate())
-        }
-    }
-}
