+++ Winey/Winey/Source/Global/Extension/UIKit/UIStackView+.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 extension UIStackView {
-    
      func addArrangedSubviews(_ views: UIView...) {
          for view in views {
              self.addArrangedSubview(view)
