+++ Winey/Winey/Source/Global/Extension/UIKit/UITextField+.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 extension UITextField {
-    
     func addLeftPadding(width: CGFloat) {
         let paddingView = UIView(frame: CGRect(x: 0, y: 0, width: width, height: self.frame.height))
         self.leftView = paddingView
