+++ Winey/Winey/Source/Global/Extension/UIKit/UIView+.swift
@@ -1,0 +1,38 @@
+//
+//  UIView+.swift
+//  Winey
+//
+//  Created by 김인영 on 2023/07/05.
+//
+
+import UIKit
+
+extension UIView {
+    func addSubviews(_ views: UIView...) {
+        views.forEach { self.addSubview($0) }
+    }
+    
+    func makeShadow(radius : CGFloat, offset: CGSize, opacity: Float) {
+        layer.shadowColor = UIColor.darkGray.cgColor
+        layer.shadowOffset = offset
+        layer.shadowRadius = radius
+        layer.shadowOpacity = opacity
+        layer.masksToBounds = false
+    }
+    
+    func makeCornerRound(radius : CGFloat) {
+        layer.cornerRadius = radius
+        layer.masksToBounds = true
+    }
+    
+    func makeBorder(width : CGFloat ,color: UIColor ) {
+        layer.borderWidth = width
+        layer.borderColor = color.cgColor
+    }
+    
+    func roundCorners(cornerRadius: CGFloat, maskedCorners: CACornerMask) {
+        clipsToBounds = true
+        layer.cornerRadius = cornerRadius
+        layer.maskedCorners = CACornerMask(arrayLiteral: maskedCorners)
+    }
+}
