+++ Winey/Winey/Source/Global/Extension/UIKit/UIViewController+.swift
@@ -8,18 +8,11 @@
 import UIKit
 
 extension UIViewController {
-    
-    /**
-     - Description: 화면 터치시 작성 종료
-     */
     /// 화면 터치시 작성 종료하는 메서드
     open override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
         self.view.endEditing(true)
     }
     
-    /**
-     - Description: 화면 터치시 키보드 내리는 Extension
-     */
     /// 화면 터치시 키보드 내리는 메서드
     func hideKeyboardWhenTappedAround() {
         let tap: UITapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(UIViewController.dismissKeyboard))
