--- Winey/Winey/Source/Global/Extension/UIView+.swift
@@ -1,65 +1,0 @@
-//
-//  UIView+.swift
-//  Winey
-//
-//  Created by 김인영 on 2023/07/05.
-//
-
-import UIKit
-
-extension UIView {
-    
-    func addSubviews(_ views: UIView...) {
-        views.forEach { self.addSubview($0) }
-    }
-    
-    func makeShadow (radius : CGFloat, offset : CGSize, opacity : Float){
-        layer.shadowColor = UIColor.darkGray.cgColor
-        layer.shadowOffset = offset
-        layer.shadowRadius = radius
-        layer.shadowOpacity = opacity
-        layer.masksToBounds = false
-    }
-    
-    func makeCornerRound (radius : CGFloat) {
-        layer.cornerRadius = radius
-        layer.masksToBounds = true
-    }
-    
-    func makeBorder (width : CGFloat ,color : UIColor ) {
-        layer.borderWidth = width
-        layer.borderColor = color.cgColor
-    }
-    
-    func setGradient() {
-        let gradient = CAGradientLayer()
-        gradient.frame = bounds
-        gradient.colors = [UIColor.init(white: 1, alpha: 0).cgColor, UIColor.init(white: 1, alpha: 1).cgColor]
-        gradient.locations = [0.0, 0.8, 1.0]
-        gradient.startPoint = CGPoint(x: 1.0, y: 0.2)
-        gradient.endPoint = CGPoint(x: 1.0, y: 1)
-        layer.insertSublayer(gradient, at: 0)
-    }
-    
-    func roundCorners(cornerRadius: CGFloat, maskedCorners: CACornerMask) {
-        clipsToBounds = true
-        layer.cornerRadius = cornerRadius
-        layer.maskedCorners = CACornerMask(arrayLiteral: maskedCorners)
-    }
-    
-    static func animateWithDamping(animation: @escaping () -> Void, completion: ((Bool) -> Void)? = nil) {
-        UIView.animate(
-            withDuration: 0.45,
-            delay: 0,
-            usingSpringWithDamping: 0.8,
-            initialSpringVelocity: 0.8,
-            options: [.allowUserInteraction],
-            animations: {
-                animation()
-            },
-            completion: { bool in
-                completion?(bool)
-            }
-        )
-    }
-}
